+++ back-end/mogul/src/main/java/com/mogul/demo/chat/nickname/NicknameGenerator.java
@@ -15,7 +15,7 @@ import java.util.*;
 @Setter
 public class NicknameGenerator {
 
-    private Map<Integer, Set<String>> nicknames;
+    private Map<Long, Set<String>> nicknames;
 
     private long prefixCount;
 
@@ -35,7 +35,7 @@ public class NicknameGenerator {
         this.random = new Random();
     }
 
-    public String generateNickname(int chatRoomId){
+    public String generateNickname(Long chatRoomId){
 
         prefixCount = nickNamePrefixRepository.count();
         postfixCount = nicknamePostfixRepository.count();
@@ -55,7 +55,7 @@ public class NicknameGenerator {
         return userId.toString();
     }
 
-    public void removeNickname(int chatRoomId, String userId) {
+    public void removeNickname(Long chatRoomId, String userId) {
         if(nicknames.containsKey(chatRoomId)){
             if(nicknames.get(chatRoomId).contains(userId)){
                 nicknames.remove(userId);
